refactor(server): renamed and moved boot files
