+++ server/src/boot/createH5PEditor.ts
@@ -1,14 +1,20 @@
-import path from 'path';
 import * as H5P from '@lumieducation/h5p-server';
 
 /**
  * Create a H5PEditor object.
  *
  * @param config the configuration object
- * @param localLibraryPath a path in the local filesystem in which the H5P libraries (content types) are stored
- * @param localContentPath a path in the local filesystem in which H5P content will be stored (only necessary if you want to use the local filesystem content storage class)
- * @param localTemporaryPath a path in the local filesystem in which temporary files will be stored (only necessary if you want to use the local filesystem temporary file storage class).
- * @param translationCallback a function that is called to retrieve translations of keys in a certain language; the keys use the i18next format (e.g. namespace:key).
+ * @param localLibraryPath a path in the local filesystem in which the H5P
+ * libraries (content types) are stored
+ * @param localContentPath a path in the local filesystem in which H5P content
+ * will be stored (only necessary if you want to use the local filesystem
+ * content storage class)
+ * @param localTemporaryPath a path in the local filesystem in which temporary
+ * files will be stored (only necessary if you want to use the local filesystem
+ * temporary file storage class).
+ * @param translationCallback a function that is called to retrieve translations
+ * of keys in a certain language; the keys use the i18next format (e.g.
+ * namespace:key).
  * @returns a H5PEditor object
  */
 export default async function createH5PEditor(
