+++ server/src/routes/updatesRoutes.ts
@@ -5,7 +5,7 @@ import * as Sentry from '@sentry/node';
 import IServerConfig from '../config/IPaths';
 import { autoUpdater, UpdateInfo, ProgressInfo } from 'electron-updater';
 
-import { io as websocket } from '../websocket';
+import { io as websocket } from '../boot/websocket';
 
 export default function (): express.Router {
     const router = express.Router();
