+++ server/src/state/electronState.ts
